Committing portfolio manager for Mostfit

+++ app/controllers/payments.rb
@@ -32,7 +32,8 @@ class Payments < Application
     @payment = Payment.get(id)
     raise NotFound unless @payment
     disallow_updation_of_verified_payments
-    
+    return unless session.user.role == :admin
+
     if @loan and @loan.delete_payment(@payment, session.user)
       do_payment(payment)
       redirect url_for_loan(@loan, 'payments'), :message => {:notice => "Payment '#{id}' has been deleted and a new one #{@payment.id} created"}
